include refactor, add simple printf tests

+++ include/mos/string.h
@@ -1,13 +1,13 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 #pragma once
 
-#include "types.h"
+#include "mos/types.h"
 
 // Inclusion of the <string.h> header may also make visible all symbols from <stddef.h>.
 #include <stddef.h>
 
 size_t strlen(const char *str);
-s8 strcmp(const char *str1, const char *str2);
+s32 strcmp(const char *str1, const char *str2);
 
 // ! The memory areas must not overlap.
 void *memcpy(void *restrict dest, const void *restrict src, size_t n);
