+++ lib/string.c
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-#include "string.h"
+#include "mos/string.h"
 
 size_t strlen(const char *str)
 {
@@ -10,7 +10,7 @@ size_t strlen(const char *str)
     return len;
 }
 
-s8 strcmp(const char *s1, const char *s2)
+s32 strcmp(const char *s1, const char *s2)
 {
     size_t i = 0;
     while (s1[i] && s2[i] && s1[i] == s2[i])
